'Refactored by Sourcery'

+++ ShasaMusic/plugins/tools/queue.py
@@ -49,36 +49,30 @@ async def ping_com(client, message: Message, _):
                 return await message.reply_text(_["cplay_4"])
             chat.title
     send = await message.reply_text(_["queue_1"])
-    if await is_active_chat(chat_id):
-        got = db.get(chat_id)
-        if got:
-            j = 0
-            msg = ""
-            for x in got:
-                j += 1
-                if j == 1:
-                    msg += f'Currently Playing:\n\n🏷Title: {x["title"]}\nDur: {x["dur"]}\nBy: {x["by"]}\n\n'
-                elif j == 2:
-                    msg += f'Queued:\n\n🏷Title: {x["title"]}\nDur: {x["dur"]}\nBy: {x["by"]}\n\n'
-                else:
-                    msg += f'🏷Title: {x["title"]}\nDur: {x["dur"]}\nBy: {x["by"]}\n\n'
-            if "Queued" in msg:
-                link = await Shasabin(msg)
-                lines = msg.count("\n")
-                if lines >= 23:
-                    car = os.linesep.join(msg.split(os.linesep)[:23])
-                else:
-                    return await send.edit_text(msg)
-                if "🏷" in car:
-                    car = car.replace("🏷", "")
-                carbon = await Carbon.generate(car, randint(100, 10000000))
-                await message.reply_photo(
-                    photo=carbon, caption=_["queue_3"].format(link)
-                )
-                await send.delete()
+    if await is_active_chat(chat_id) and (got := db.get(chat_id)):
+        msg = ""
+        for j, x in enumerate(got, start=1):
+            if j == 1:
+                msg += f'Currently Playing:\n\n🏷Title: {x["title"]}\nDur: {x["dur"]}\nBy: {x["by"]}\n\n'
+            elif j == 2:
+                msg += f'Queued:\n\n🏷Title: {x["title"]}\nDur: {x["dur"]}\nBy: {x["by"]}\n\n'
             else:
-                await send.edit_text(msg)
+                msg += f'🏷Title: {x["title"]}\nDur: {x["dur"]}\nBy: {x["by"]}\n\n'
+        if "Queued" in msg:
+            link = await Shasabin(msg)
+            lines = msg.count("\n")
+            if lines >= 23:
+                car = os.linesep.join(msg.split(os.linesep)[:23])
+            else:
+                return await send.edit_text(msg)
+            if "🏷" in car:
+                car = car.replace("🏷", "")
+            carbon = await Carbon.generate(car, randint(100, 10000000))
+            await message.reply_photo(
+                photo=carbon, caption=_["queue_3"].format(link)
+            )
+            await send.delete()
         else:
-            await send.edit_text(_["queue_2"])
+            await send.edit_text(msg)
     else:
         await send.edit_text(_["queue_2"])
